+++ staging/src/k8s.io/apiextensions-apiserver/pkg/controller/crdregistration/crdregistration_controller.go
